requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
sentence-transformers>=2.3.1
openai>=1.10.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
streamlit>=1.30.0
arxiv>=2.1.0
pypdf2>=3.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
pytest>=7.4.0
black>=24.1.0
ruff>=0.1.0
pre-commit>=3.6.0
